feat: board 및 common components 쪼개기
