+++ itta-bab-frontend/src/components/common/SearchBarAndSort.vue
@@ -1,0 +1,50 @@
+<!-- SearchBarAndSort.vue -->
+<template>
+  <div class="search-sort-container">
+    <div class="search-bar">
+      <input type="text" placeholder="원하는 내용을 작성해서 검색하세요." />
+    </div>
+    <div class="sort-dropdown">
+      <select>
+        <option value="latest">최신순</option>
+        <option value="mostHated">좋아요 적은순</option>
+        <option value="mostLiked">좋아요 많은 순</option>
+      </select>
+    </div>
+  </div>
+</template>
+
+<style scoped>
+.search-sort-container {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  gap: 80px;
+  width: 90%;
+}
+.search-bar {
+  display: flex;
+  align-items: center;
+  background-color: white;
+  border-radius: 25px;
+  padding: 10px 15px;
+  width: 70%;
+  border: 1px solid var(--background-color);
+  flex: 1;
+}
+.search-bar input {
+  flex: 1;
+  border: none;
+  outline: none;
+  font-size: 1em;
+}
+.sort-dropdown {
+  flex-shrink: 0;
+}
+.sort-dropdown select {
+  padding: 10px;
+  border-radius: 12px;
+  border: 1px solid var(--white);
+  background-color: var(--white);
+}
+</style>
